PR comments cleanup
