+++ GW2EIEvtcParser/EncounterLogic/EncounterIDs.cs
@@ -38,7 +38,7 @@ internal static class EncounterIDs
         public const long SilentSurfMask = 0x000400;
         public const long LonelyTowerMask = 0x000500;
         public const long CaptainMaiTrinBossMask = 0x000700;
-        public const long DeepstoneMask = 0x000700;
+        public const long DeepstoneMask = 0x000800;
     }
 
     public static class StrikeMasks
